[Gold II] Title: 나누기, Time: 260 ms, Memory: 24596 KB, Score: 65 point -BaekjoonHub

+++ 백준/Gold/21757. 나누기/나누기.java
@@ -50,7 +50,7 @@ class Main {
 
 
         if(prefixSum[n] == 0){
-            int cnt = 0;
+            long cnt = 0;
             for(int i=1 ; i<prefixSum.length ; i++){
                 if(prefixSum[i] == 0)
                     cnt++;
